Refactored controllers nook_statistics and location_statistics
