+++ config/application.rb
@@ -16,6 +16,9 @@ module Bookanook
     # Run "rake -D time" for a list of tasks for finding time zone names. Default is UTC.
     config.time_zone = 'Eastern Time (US & Canada)'
 
+    # Include the lib/ path to allow files
+    config.autoload_paths += %W(#{config.root}/lib)
+
     # The default locale is :en and all translations from config/locales/*.rb,yml are auto loaded.
     # config.i18n.load_path += Dir[Rails.root.join('my', 'locales', '*.{rb,yml}').to_s]
     # config.i18n.default_locale = :de
